🚀feat: xml support

+++ jar-cart/downloader.js
@@ -7,8 +7,20 @@ const path = require("node:path");
 async function downloadJars(manifestDeps, libDir) {
   const requiredFiles = new Map();
 
+  function getRawList(deps) {
+    if (!deps) return [];
+    if (Array.isArray(deps)) return deps;
+    if (deps.dependency) {
+      return Array.isArray(deps.dependency)
+        ? deps.dependency
+        : [deps.dependency];
+    }
+    return [];
+  }
+
   function flatten(list) {
-    for (const d of list) {
+    const items = getRawList(list);
+    for (const d of items) {
       const fileName = `${d.library}-${d.version}.jar`;
       if (!requiredFiles.has(fileName)) {
         requiredFiles.set(fileName, {
@@ -18,7 +30,8 @@ async function downloadJars(manifestDeps, libDir) {
           fileName: fileName,
         });
       }
-      if (d.dependencies && d.dependencies.length > 0) {
+
+      if (d.dependencies) {
         flatten(d.dependencies);
       }
     }
@@ -26,11 +39,15 @@ async function downloadJars(manifestDeps, libDir) {
 
   flatten(manifestDeps);
 
-  const existingFiles = await fs.readdir(libDir);
-  for (const file of existingFiles) {
-    if (file.endsWith(".jar") && !requiredFiles.has(file)) {
-      await fs.remove(path.join(libDir, file));
+  if (await fs.pathExists(libDir)) {
+    const existingFiles = await fs.readdir(libDir);
+    for (const file of existingFiles) {
+      if (file.endsWith(".jar") && !requiredFiles.has(file)) {
+        await fs.remove(path.join(libDir, file));
+      }
     }
+  } else {
+    await fs.ensureDir(libDir);
   }
 
   for (const dep of requiredFiles.values()) {
@@ -44,21 +61,21 @@ async function downloadJars(manifestDeps, libDir) {
       const response = await axios({
         url,
         responseType: "stream",
-        headers: { "User-Agent": "JarCart-VSCode/1.0.3" },
+        headers: { "User-Agent": "JarCart-VSCode/1.1.1" },
       });
 
       const writer = fs.createWriteStream(filePath);
       response.data.pipe(writer);
 
-      await new Promise((resolve) => {
+      await new Promise((resolve, reject) => {
         writer.on("finish", resolve);
-        writer.on("error", () => {
+        writer.on("error", (err) => {
           writer.close();
-          resolve();
+          reject(err);
         });
       });
     } catch (e) {
-      vscode.window.showErrorMessage(`Failed to download ${dep.fileName}`);
+      vscode.window.showErrorMessage(`Failed to download ${dep.fileName} ❌`);
     }
   }
 }
